Migrated synchblockchain test to jest
